Commit for mini website project for ZTM.
This is for the ZTM Complete Web Developer Course. I'd thought I'd
save these changes in the cloud, so I can preserve this knowledge for
when I need it. For example, to preserve an image's aspect ratio while
making it fit into your website, use max-width and max-height, not
width and height.